Create Sql project
This project represents the database design for a comprehensive online retail application. The schema incorporates several tables and attributes to simulate the complex operations typical of an e-commerce platform. The purpose of this database is to manage user interactions, orders, payments, inventory, and other critical components of an online shopping experience.